Fix stack alignment for JIT function call targets
X64_84 ABIs demand that the stack pointer is a multiple of 16 before
CALLing a function. Compilers such as GCC create optimized code
based on this constraint, e.g. code with MOVAPS in which the operand
must be aligned on a 16-byte boundary.

The dynamips JIT does not ensure alignment before CALLs. This led to
SIGSEGV crashes of dynamips in libc on Ubuntu 22.04 and earlier.

There are multiple options to deal with this:
    1. Fix in the caller: The JIT is changed to generate code that
                          ensures proper stack pointer alignment
    2. Fix in the callee: Functions called from JIT generated code
                          ensure proper stack pointer alignment

Option 2 is implemented in this change by adding the
	force_align_arg_pointer
attribute to functions called from JIT generated code.