Fix uninitialized field in ppc405_slow_lookup
The field mts_map_t.offset was uninitialized.
This causes ppc32_mem_map to get unpredictable memory.

This code is only used when the pvr register is PPC32_PVR_405.
Only ppc32_vmtest_init_platform sets that and the line is commented out.
Therefore no platforms are affected and it was still in development.