Ensure the full vtty replay buffer is sent.
send is allowed to send only part of the buffer.
Assuming it sends everything could cause information loss.

A non-full replay buffer was detected with nul bytes.
This means valid nul bytes could be skipped while sending the replay buffer.
Now it remembers if the buffer is full with vtty->replay_full.